requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
websockets>=11.0,<12.0

# Blockchain & Crypto
web3==6.11.0
solana>=0.34.0
solders>=0.20.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0

# Scheduling & Async
celery>=5.3.0
apscheduler>=3.10.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
ta>=0.10.0  # Technical Analysis

# Machine Learning (optional)
scikit-learn>=1.3.0

# Notifications
python-telegram-bot>=20.7
discord.py>=2.3.0

# Web Scraping
beautifulsoup4>=4.12.0
selenium>=4.15.0
playwright>=1.40.0

# API Clients
ccxt>=4.0.0  # Crypto exchange library
pycoingecko>=3.1.0

# Monitoring & Logging
loguru>=0.7.0
sentry-sdk>=1.39.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# OpenAI for AI analysis (optional)
openai>=1.3.0
